8
Started with making mobs. A new mob class is made, and the spawn mechanic that loads the map now loads a mob whereever it sees "M" within the file.

Direction vectors:
To enable simple mob tracking of the player sprite, vectors become extremely useful as they allow me to easily calculate the angle of depression of the mob from the x axis to calculate how much the mob needs to move. 

From the top left, there is a vector pointing towards the mob, and there is also a vector pointing towards the player. The mobs vector is mob.pos and the players vector is player.pos

To make the mob always point towards the player, I minused mob.pos from player.pos to get the angle of depression/elevation from the mob to the player, so the mob knows where the player sprite is at all times and can track the player.

To get the angle, we need to look at the angle in relation to the x axis, which in vector terms is actually vec(1,0). Thus, we need to find the angle between the x vector and the vector between the mob and player

PUTTING THIS ALL TOGETHER into code:
self.rot = (self.game.player.pos - self.pos).angle_to(vec(1,0))

where self is the mob object and game is referenced within the mob class.

I also added a small graphic for the mob.