9
Added alot.

Made mobs have a simple AI. They know the angle where they should point towards the player, and now move towards the player.
I encountered many problems while doing this, specifically the movement of the mob. The mob had no friction, thus it was accelerating at a constant rate, which made the mobs go faster and faster. Thus, I used 
self.acc += self.vel * -1
which means that the faster the mob is going, the higher this number will be, so the less he will accelerate. By setting a max speed by this, it will still accelerate, however will cap the speed and slow mobs down when the mobs get fast.

Now, mob collision with walls must be taken into account.

I removed collide_with_walls as a method for the player class, and modified it so every sprite that I call into the function can use the function to detect whether it has collide with any walls.

Thus, mobs can now collide with walls

I added the ability to shoot. I made a bullet class, and made the bullet have many attributes, such as bullet lifetime, bullet speed, bullet rate, and rate of fire. I also implemented a gun spread attribute, which enables slightly randomness when bullets are being shot so they aren't just a boring straight line of bullets. (recoil)

I also added knockback to the player when the player presses space to shoot. This enables a more realistic experience, so the player moves slightly back when the gun gets shot.

After that I added stopping power for the gun, so when bullets hit mobs, the bullets will have stopping power so the bullets will slow down the mobs in their tracks, again making for more realistic and enjoyable gameplay.

I added health bars on the mobs, which go down and change color depending on how much health the mobs have
If the mob has over 60% health, the health bar will be green
elif over 30%, the health bar will be yellow,
elif under 30%, the health bar will be red.

I also added a health bar on the top left, as part of the games HUD (heads up display), which displays the players health in a similar fashion that it displays the mobs health.

Next, I will improve on the map and explore the usage of a professionalised map editor in order to assist in the development of a more sophisticated mapping system.