10
This commit is huge.

Just some definitions to ensure that readers know what I'm talking about.
Tiled - open map tilemap editor, really powerful and supported by many game engines. Python has a library called pytmx which is specifically made to support tilemaps, and thus I will be using this to support my tilemap.

I will be using kenney.nl's top down tilemap graphics pack in order to add graphics to my game. 

Tiled is extremely great as kenney provides a spritesheet where I can divide each seperate tile on the spritesheet into a small picture, and use those pictures to essentially paint my map.
Tiled also supports the tagging of different elements within the map, thus this allows pytmx to see what it should do with the tile that it is rendering. It is similar to the previous text based map where I used code to parse through every character in the text file, and check what that character was, and if that character was a M, I spawned a mob in that position, if it was a P, it would spawn a player in that position etc..

Tiled is essentially an image editor which is specifically tailored to making levels and maps for games. It has layers, just like in photoshop, and I used those layers to have a background layer of grass, and then layers on top of that grass, such as trees and fences.

Then, I created objects in Tiled, with the tag "wall" 

I then coded in another class called Obstacle. This class is essentially the Wall class however made for tilemaps. This makes the Wall class obsolete. I left it in just in case I require it for future debugging purposes or if I ever want to revert back to text files. (which is never)

Just on a side note, the reason why text files suck soo much for my purposes for this game, is that I have many many pictures that I want to use from kenney.nl's art pack. If I was to use text files to do this, it would create many problems. I would have to use a different character to indicate what I want to be shown on the map in the text file (e.g. G for grass, P for pole), and then manually render each one onto the map. 

With Tiled, it allows me to graphically do this, and creates a more efficient method as pytmx is a great library that fully supports Tiled.

So within Tiled, I added obstacles to my map. The attributes of these elements of the map are made into a dictionary by pytmx. So I can parse through the dictionary, and for every element of the map (tile in the map) that is tagged as an obstacle, I used the code from my previous "Wall" class, and thus created walls through that. 

Tiled also allows me to create custom hitboxes on whatever I want, and of differing size.

Within my code, I have improved the implementation of my DEBUG_MODE, to enable it to show all hitboxes of all sprites of the game, when set to "ON". 

After implementation of all these features, working with pytmx and the implementation of my tilemap made in Tiled, and creating the new code needed to adapt these changes (documentation of which is written within the code as usual), I added zombies to the map through tiled, and similar to the previous text based map mechanic, I scanned for every tile in the map to detect a zombie and spawn a zombie in its respective location.

Now moving on to the final edits of this commit:
There is a problem with the zombies "AI" movement. It was very static, they moved in a straight line and the movement was extremely predictable. Because they move in a linear direction, the zombies will eventually clump up into a bunch and look like one single zombie. I needed to diversify the movement of the zombies to make the movements of these zombies varied and more realistic.

I made some edits to how they move, and added a certain radius where they were not allowed to be next to another zombie. Basically each zombie now had its own personal space, which means they would not clump together. This also meant that they would begin circling the player sprite, and thus make a more challenging game. The code is documented as usual.

After all this, I also added a random list of zombie speeds to make that more interesting, instead of having all the zombies move at a constant speed. Props to the random.choice function from the random library for making this easier.

// end